requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic
alembic
pytest-asyncio
httpx
sqlalchemy[asyncio]
pydantic[email]
pytest
pytest-timeout
pytest-mock
asgi-lifespan
bcrypt==4.0.1